python/requirements.txt
furhat-realtime-api
python-dotenv
streamlit>=1.31.0
httpx>=0.27.0
asyncio
uvloop
